rename: 리디자인, 이전 적용 디자인 관련 포트폴리오 아이템 컴포넌트명 변경, 공통 스타일링 코드 파일명 변경
