+++ src/pages/MyPortfolio.tsx
@@ -5,10 +5,9 @@ import { FaPlus } from 'react-icons/fa';
 import { getMyPortfolio } from '@src/apis/portfolio';
 import { PortfolioDataType } from '@src/types/portfolioType';
 import { PATH_URL } from '@src/constants/constants';
-import * as S from '@src/style/common/mainPagePortfolioStyle';
-import PortfolioItem from '@src/components/main/PortfolioItem';
+import * as S from '@src/style/common/commonStyles';
 import NoPortfolio from '@src/components/myPortfolio/NoPortfolio';
-import TestReDesignPortfolioItem from '@src/components/common/TestReDesignPortfolioItem';
+import PortfolioItem from '@src/components/common/PortfolioItem';
 
 const MyPortfolio = () => {
   const [myPortfolioList, setMyPortfolioList] = useState<PortfolioDataType[]>([]);
@@ -44,7 +43,7 @@ const MyPortfolio = () => {
               <StCreateText>새 포트폴리오 작성</StCreateText>
             </StCreatePortfolio>
             {myPortfolioList?.map(portfolio => (
-              <TestReDesignPortfolioItem key={portfolio.id} item={portfolio} />
+              <PortfolioItem key={portfolio.id} item={portfolio} />
             ))}
           </S.PortfolioListContainer>
         </StMyPortfolioContainer>
